Adding "NumberOfAltPlace"
"NumberOfAltPlace" and "ProperDateInterval" will work together to
further split a span

For example, if "NumberOfAltPlace" is 3, "ProperDateInterval" is 30. A
span is equally splited to 3 alternative y position. By default, the
spot is pointed to the middle, if the date interval exceed 30 days, it
will comes to 1st alternative y, if the date interval on that position
also exceed 30 days, it will comes to 2nd alternative y, etc

If Current Setting don't permit all event to be displayed. A error
message will show up

---
No place to put the new stuff ...
---

Consider to either increase the "NumberOfAltPlace" or decrease
"ProperDateInterval" to resolve this issue

Signed-off-by: ericsunplus <ericsunplus@gmail.com>